systemd-journal: extract mock specific test cases to separate unit test
